requirements.txt
streamlit==1.32.1
openai==1.3.0
python-dotenv==1.0.0
pandas==2.1.3
matplotlib==3.8.2
numpy==1.26.2
plotly>=5.18.0
st-gsheets-connection
